+++ zed_components/src/zed_camera/src/zed_camera_component_main.cpp
@@ -287,57 +287,81 @@ void ZedCamera::initServices()
     mResetOdomSrv = create_service<std_srvs::srv::Trigger>(
       srv_name,
       std::bind(&ZedCamera::callback_resetOdometry, this, _1, _2, _3));
-    RCLCPP_INFO(get_logger(), " * '%s'", mResetOdomSrv->get_service_name());
+    RCLCPP_INFO_STREAM(
+      get_logger(), " * Advertised on service: '" << mResetOdomSrv->get_service_name() << "'");
     // Reset Pose
     srv_name = srv_prefix + mSrvResetPoseName;
     mResetPosTrkSrv = create_service<std_srvs::srv::Trigger>(
       srv_name,
       std::bind(&ZedCamera::callback_resetPosTracking, this, _1, _2, _3));
-    RCLCPP_INFO(get_logger(), " * '%s'", mResetPosTrkSrv->get_service_name());
+    RCLCPP_INFO_STREAM(
+      get_logger(), " * Advertised on service: '" << mResetPosTrkSrv->get_service_name() << "'");
     // Set Pose
     srv_name = srv_prefix + mSrvSetPoseName;
     mSetPoseSrv = create_service<zed_msgs::srv::SetPose>(
       srv_name, std::bind(&ZedCamera::callback_setPose, this, _1, _2, _3));
-    RCLCPP_INFO(get_logger(), " * '%s'", mSetPoseSrv->get_service_name());
+    RCLCPP_INFO_STREAM(
+      get_logger(),
+      " * Advertised on service: '" << mSetPoseSrv->get_service_name() << "'");
+    // Save Area Memory
+    srv_name = srv_prefix + mSrvSaveAreaMemoryName;
+    mSaveAreaMemorySrv = create_service<zed_msgs::srv::SetROI>(
+      // TODO(Walter) change with "zed_msgs::srv::SetAreaMemory" when available via APT
+      srv_name, std::bind(&ZedCamera::callback_saveAreaMemory, this, _1, _2, _3));
+    RCLCPP_INFO_STREAM(
+      get_logger(),
+      " * Advertised on service: '" << mSaveAreaMemorySrv->get_service_name() << "'");
 
     // Enable Object Detection
     srv_name = srv_prefix + mSrvEnableObjDetName;
     mEnableObjDetSrv = create_service<std_srvs::srv::SetBool>(
       srv_name,
       std::bind(&ZedCamera::callback_enableObjDet, this, _1, _2, _3));
-    RCLCPP_INFO(get_logger(), " * '%s'", mEnableObjDetSrv->get_service_name());
+    RCLCPP_INFO_STREAM(
+      get_logger(), " * Advertised on service: '" << mEnableObjDetSrv->get_service_name() << "'");
 
     // Enable BodyTracking
     srv_name = srv_prefix + mSrvEnableBodyTrkName;
     mEnableBodyTrkSrv = create_service<std_srvs::srv::SetBool>(
       srv_name,
       std::bind(&ZedCamera::callback_enableBodyTrk, this, _1, _2, _3));
-    RCLCPP_INFO(get_logger(), " * '%s'", mEnableBodyTrkSrv->get_service_name());
+    RCLCPP_INFO_STREAM(
+      get_logger(),
+      " * Advertised on service: '" << mEnableBodyTrkSrv->get_service_name() << "'");
 
     // Enable Mapping
     srv_name = srv_prefix + mSrvEnableMappingName;
     mEnableMappingSrv = create_service<std_srvs::srv::SetBool>(
       srv_name,
       std::bind(&ZedCamera::callback_enableMapping, this, _1, _2, _3));
-    RCLCPP_INFO(get_logger(), " * '%s'", mEnableMappingSrv->get_service_name());
+    RCLCPP_INFO_STREAM(
+      get_logger(),
+      " * Advertised on service: '" << mEnableMappingSrv->get_service_name() << "'");
+
+
   }
 
   // Enable Streaming
   srv_name = srv_prefix + mSrvEnableStreamingName;
   mEnableStreamingSrv = create_service<std_srvs::srv::SetBool>(
     srv_name, std::bind(&ZedCamera::callback_enableStreaming, this, _1, _2, _3));
-  RCLCPP_INFO(get_logger(), " * '%s'", mEnableStreamingSrv->get_service_name());
+  RCLCPP_INFO_STREAM(
+    get_logger(),
+    " * Advertised on service: '" << mEnableStreamingSrv->get_service_name() << "'");
 
   // Start SVO Recording
   srv_name = srv_prefix + mSrvStartSvoRecName;
   mStartSvoRecSrv = create_service<zed_msgs::srv::StartSvoRec>(
     srv_name, std::bind(&ZedCamera::callback_startSvoRec, this, _1, _2, _3));
-  RCLCPP_INFO(get_logger(), " * '%s'", mStartSvoRecSrv->get_service_name());
+  RCLCPP_INFO_STREAM(
+    get_logger(), " * Advertised on service: '" << mStartSvoRecSrv->get_service_name() << "'");
   // Stop SVO Recording
   srv_name = srv_prefix + mSrvStopSvoRecName;
   mStopSvoRecSrv = create_service<std_srvs::srv::Trigger>(
     srv_name, std::bind(&ZedCamera::callback_stopSvoRec, this, _1, _2, _3));
-  RCLCPP_INFO(get_logger(), " * '%s'", mStopSvoRecSrv->get_service_name());
+  RCLCPP_INFO_STREAM(
+    get_logger(),
+    " * Advertised on service: '" << mStopSvoRecSrv->get_service_name() << "'");
 
   // Pause SVO (only if the realtime playing mode is disabled)
   if (mSvoMode) {
@@ -348,7 +372,9 @@ void ZedCamera::initServices()
     mPauseSvoSrv = create_service<std_srvs::srv::Trigger>(
       srv_name,
       std::bind(&ZedCamera::callback_pauseSvoInput, this, _1, _2, _3));
-    RCLCPP_INFO(get_logger(), " * '%s'", mPauseSvoSrv->get_service_name());
+    RCLCPP_INFO_STREAM(
+      get_logger(),
+      " * Advertised on service: '" << mPauseSvoSrv->get_service_name() << "'");
 #ifndef USE_SVO_REALTIME_PAUSE
   }
 #endif
@@ -358,31 +384,40 @@ void ZedCamera::initServices()
     mSetSvoFrameSrv = create_service<zed_msgs::srv::SetSvoFrame>(
       srv_name,
       std::bind(&ZedCamera::callback_setSvoFrame, this, _1, _2, _3));
-    RCLCPP_INFO(get_logger(), " * '%s'", mSetSvoFrameSrv->get_service_name());
+    RCLCPP_INFO_STREAM(
+      get_logger(), " * Advertised on service: '" << mSetSvoFrameSrv->get_service_name() << "'");
   }
 
   // Set ROI
   srv_name = srv_prefix + mSrvSetRoiName;
   mSetRoiSrv = create_service<zed_msgs::srv::SetROI>(
     srv_name, std::bind(&ZedCamera::callback_setRoi, this, _1, _2, _3));
-  RCLCPP_INFO(get_logger(), " * '%s'", mSetRoiSrv->get_service_name());
+  RCLCPP_INFO_STREAM(
+    get_logger(),
+    " * Advertised on service: '" << mSetRoiSrv->get_service_name() << "'");
   // Reset ROI
   srv_name = srv_prefix + mSrvResetRoiName;
   mResetRoiSrv = create_service<std_srvs::srv::Trigger>(
     srv_name, std::bind(&ZedCamera::callback_resetRoi, this, _1, _2, _3));
-  RCLCPP_INFO(get_logger(), " * '%s'", mResetRoiSrv->get_service_name());
+  RCLCPP_INFO_STREAM(
+    get_logger(),
+    " * Advertised on service: '" << mResetRoiSrv->get_service_name() << "'");
 
   if (mGnssFusionEnabled) {
     // To Latitude/Longitude
     srv_name = srv_prefix + mSrvToLlName;
     mToLlSrv = create_service<robot_localization::srv::ToLL>(
       srv_name, std::bind(&ZedCamera::callback_toLL, this, _1, _2, _3));
-    RCLCPP_INFO(get_logger(), " * '%s'", mToLlSrv->get_service_name());
+    RCLCPP_INFO_STREAM(
+      get_logger(),
+      " * Advertised on service: '" << mToLlSrv->get_service_name() << "'");
     // From Latitude/Longitude
     srv_name = srv_prefix + mSrvFromLlName;
     mFromLlSrv = create_service<robot_localization::srv::FromLL>(
       srv_name, std::bind(&ZedCamera::callback_fromLL, this, _1, _2, _3));
-    RCLCPP_INFO(get_logger(), " * '%s'", mFromLlSrv->get_service_name());
+    RCLCPP_INFO_STREAM(
+      get_logger(),
+      " * Advertised on service: '" << mFromLlSrv->get_service_name() << "'");
   }
 }
 
@@ -1576,9 +1611,13 @@ void ZedCamera::getPosTrackingParams()
       shared_from_this(), "pos_tracking.area_memory",
       mAreaMemory, mAreaMemory, " * Area Memory: ");
     sl_tools::getParam(
-      shared_from_this(), "pos_tracking.area_memory_db_path",
+      shared_from_this(), "pos_tracking.area_file_path",
       mAreaMemoryDbPath, mAreaMemoryDbPath,
       " * Area Memory DB: ");
+    sl_tools::getParam(
+      shared_from_this(), "pos_tracking.save_area_memory_on_closing",
+      mSaveAreaMemoryOnClosing, mSaveAreaMemoryOnClosing,
+      " * Save Area Memory on closing: ");
   }
   sl_tools::getParam(
     shared_from_this(), "pos_tracking.set_as_static",
@@ -4214,6 +4253,16 @@ void ZedCamera::closeCamera()
 
   RCLCPP_INFO(get_logger(), "=== CLOSING CAMERA ===");
 
+  if( mPosTrackingStarted && !mAreaMemoryDbPath.empty() && mSaveAreaMemoryOnClosing ) {
+    sl::ERROR_CODE err = mZed->saveAreaMap(mAreaMemoryDbPath.c_str();
+    if(r!=sl::ERROR_CODE::SUCCESS)
+    {
+      RCLCPP_STREAM_WARN(get_logger(), "Error saving the Area Memory file: " << sl::toString(err) << " - " << sl::toVerbose(err)); 
+    } else {
+      RCLCPP_STREAM_INFO(get_logger(), "Area Memory file saved: " << mAreaMemoryDbPath );
+    }
+  }
+
   mZed->close();
   mZed.reset();
   DEBUG_COMM("Camera closed");
@@ -4387,7 +4436,7 @@ bool ZedCamera::startPosTracking()
     mAreaMemoryDbPath = "";
     RCLCPP_WARN_STREAM(
       get_logger(),
-      "'area_memory_db_path' path doesn't exist or is unreachable: "
+      "'area_file_path' path doesn't exist or is unreachable: "
         << mAreaMemoryDbPath);
   }
 
@@ -8695,6 +8744,14 @@ void ZedCamera::callback_pubPaths()
       DEBUG_STREAM_COMM("Message publishing generic ecception: ");
     }
   }
+}
+
+void ZedCamera::callback_saveAreaMemory(
+  const std::shared_ptr<rmw_request_id_t> request_header,
+  const std::shared_ptr<zed_msgs::srv::SetROI_Request> req,
+  std::shared_ptr<zed_msgs::srv::SetROI_Response> res)
+{
+
 }
 
 void ZedCamera::callback_resetOdometry(